Safari-609.1.11.1/LayoutTests/media/session/releasing-sessions-expected.txt

Sessions that are explicitly released should have their media elements indefinitely paused and go into an Idle state.


EVENT(canplaythrough)
EXPECTED (video.paused == 'true') OK
Begin playing media.
EVENT(playing)
Media began playing.
EXPECTED (video.paused == 'false') OK
EXPECTED (internals.mediaSessionCurrentState(session) == 'active') OK
RUN(session.deactivate())
EXPECTED (video.paused == 'true') OK
EXPECTED (internals.mediaSessionCurrentState(session) == 'idle') OK
END OF TEST
